This is an experimental branch
The initial idea was to resolve a conflict between RTWYSIWYG and RTWIKI
where users would have difficulty joining the RTWiki session.

Several other issues were added to the work while it was in progress and
as a result the commit became rather large. For other developers' sanity
it will be broken into several commits with the relevant JIRA issues.